# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * OriginLock API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface UserResponse {
  id: string;
  email: string;
  emailVerified: boolean;
  createdAt: string;
  hasProfile: boolean;
}

export interface AuthResponse {
  user: UserResponse;
  token: string;
}

export interface SignupBody {
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export interface ForgotPasswordBody {
  email: string;
}

export interface ResetPasswordBody {
  token: string;
  /** @minLength 8 */
  password: string;
}

export interface CreatorProfileResponse {
  id: string;
  userId: string;
  displayName: string;
  creatorHandle: string;
  /** @nullable */
  legalName?: string | null;
  /** @nullable */
  pseudonym?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  claimStatement?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface CreateProfileBody {
  displayName: string;
  creatorHandle: string;
  /** @nullable */
  legalName?: string | null;
  /** @nullable */
  pseudonym?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  claimStatement?: string | null;
}

export interface UpdateProfileBody {
  displayName?: string;
  creatorHandle?: string;
  /** @nullable */
  legalName?: string | null;
  /** @nullable */
  pseudonym?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  claimStatement?: string | null;
}

export interface ProjectResponse {
  id: string;
  userId: string;
  title: string;
  /** @nullable */
  description?: string | null;
  fileCount: number;
  createdAt: string;
  updatedAt: string;
}

export type FileRecordResponsePrivacyMode =
  (typeof FileRecordResponsePrivacyMode)[keyof typeof FileRecordResponsePrivacyMode];

export const FileRecordResponsePrivacyMode = {
  private: "private",
  shareable_certificate: "shareable_certificate",
  public_verification: "public_verification",
} as const;

export interface FileRecordResponse {
  id: string;
  userId: string;
  /** @nullable */
  projectId?: string | null;
  /** @nullable */
  projectTitle?: string | null;
  title: string;
  originalFilename: string;
  fileType: string;
  mimeType: string;
  fileSizeBytes: string;
  sha256Hash: string;
  /** @nullable */
  notes?: string | null;
  privacyMode: FileRecordResponsePrivacyMode;
  /** @nullable */
  certificateId?: string | null;
  /** @nullable */
  certificateStatus?: string | null;
  /** @nullable */
  verificationUrl?: string | null;
  /** @nullable */
  recordedAtUtc?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectDetailResponse {
  id: string;
  userId: string;
  title: string;
  /** @nullable */
  description?: string | null;
  fileCount: number;
  files: FileRecordResponse[];
  createdAt: string;
  updatedAt: string;
}

export interface CreateProjectBody {
  title: string;
  /** @nullable */
  description?: string | null;
}

export interface UpdateProjectBody {
  title?: string;
  /** @nullable */
  description?: string | null;
}

export type FileRecordDetailResponsePrivacyMode =
  (typeof FileRecordDetailResponsePrivacyMode)[keyof typeof FileRecordDetailResponsePrivacyMode];

export const FileRecordDetailResponsePrivacyMode = {
  private: "private",
  shareable_certificate: "shareable_certificate",
  public_verification: "public_verification",
} as const;

export interface FileRecordDetailResponse {
  id: string;
  userId: string;
  /** @nullable */
  projectId?: string | null;
  /** @nullable */
  projectTitle?: string | null;
  creatorProfileId: string;
  displayName: string;
  creatorHandle: string;
  title: string;
  originalFilename: string;
  fileType: string;
  mimeType: string;
  fileSizeBytes: string;
  sha256Hash: string;
  /** @nullable */
  notes?: string | null;
  privacyMode: FileRecordDetailResponsePrivacyMode;
  /** @nullable */
  certificateId?: string | null;
  /** @nullable */
  certificateStatus?: string | null;
  /** @nullable */
  verificationUrl?: string | null;
  /** @nullable */
  recordedAtUtc?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type UpdateFileBodyPrivacyMode =
  (typeof UpdateFileBodyPrivacyMode)[keyof typeof UpdateFileBodyPrivacyMode];

export const UpdateFileBodyPrivacyMode = {
  private: "private",
  shareable_certificate: "shareable_certificate",
  public_verification: "public_verification",
} as const;

export interface UpdateFileBody {
  title?: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  projectId?: string | null;
  privacyMode?: UpdateFileBodyPrivacyMode;
}

export interface UploadResponse {
  fileId: string;
  certificateId: string;
  verificationUrl: string;
  sha256Hash: string;
  recordedAtUtc: string;
}

export type VerificationResponseStatus =
  (typeof VerificationResponseStatus)[keyof typeof VerificationResponseStatus];

export const VerificationResponseStatus = {
  valid: "valid",
  revoked: "revoked",
  superseded: "superseded",
} as const;

/**
 * @nullable
 */
export type VerificationResponseHashMatchResult =
  | (typeof VerificationResponseHashMatchResult)[keyof typeof VerificationResponseHashMatchResult]
  | null;

export const VerificationResponseHashMatchResult = {
  match: "match",
  mismatch: "mismatch",
} as const;

export interface VerificationResponse {
  certificateId: string;
  status: VerificationResponseStatus;
  displayName: string;
  creatorHandle: string;
  /** @nullable */
  projectTitle?: string | null;
  fileTitle: string;
  originalFilename: string;
  fileType: string;
  mimeType: string;
  fileSizeBytes: string;
  sha256Hash: string;
  recordedAtUtc: string;
  /** @nullable */
  hashMatchResult?: VerificationResponseHashMatchResult;
}

export interface DashboardStatsResponse {
  protectedFiles: number;
  projects: number;
  certificates: number;
  verificationChecks: number;
}

export type UploadFileBodyPrivacyMode =
  (typeof UploadFileBodyPrivacyMode)[keyof typeof UploadFileBodyPrivacyMode];

export const UploadFileBodyPrivacyMode = {
  private: "private",
  shareable_certificate: "shareable_certificate",
  public_verification: "public_verification",
} as const;

export type UploadFileBody = {
  file: Blob;
  title: string;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  projectId?: string | null;
  privacyMode?: UploadFileBodyPrivacyMode;
};

export type VerifyCertificateParams = {
  submittedHash?: string;
};